+++ arm_inlines.h
@@ -164,25 +164,61 @@ __INLINE LWORD conv32upto64Bits(const DWORD val, const BYTE nb) {
 	return ((LWORD) ((LWORD) (val << nb) + (LWORD) (val & (0xFFFFFFFFUL >> (32-nb))))); }
 
 
+/*************************/
+/***  ENDIANS TESTING  ***/
+/*************************/
+
+/*!\brief Test Core endian
+** \return Endian type
+**/
+__INLINE eEndian testEndian_basic(void) {
+	WORD x = 0x100;
+	return (* (CHAR *) (&x)) ? Endian_big : Endian_little; }
+
+/*!\brief Test Core endian (full, recognizing mid endians too)
+** \return Endian type
+**/
+__INLINE eEndian testEndian_full(void) {
+	union {
+		LWORD	dword;
+		BYTE	byte[sizeof(LWORD)];
+	} tst;
+
+	tst.byte[0] = 1;
+	tst.byte[1] = 2;
+	tst.byte[2] = 3;
+	tst.byte[3] = 4;
+
+	switch (tst.dword)
+	{
+		case 0x01020304UL:	return Endian_big;
+		case 0x04030201UL:	return Endian_little;
+		case 0x03040102UL:	return Endian_mid_big;
+		case 0x02010403UL:	return Endian_mid_little;
+		default:      		return Endian_unknown;
+	}
+}
+
+
 /*****************************/
 /***  ENDIANS CONVERSIONS  ***/
 /*****************************/
 
-/*!\brief Swap endians of the contents of a 16b value
+/*!\brief Swap endians of the contents of a 16b value (little -> big, big -> little)
 ** \param[in] w - 16b value
 ** \return Swapped value
 **/
 __INLINE WORD SWAP_END16B(const WORD w) {
 	return (WORD) (LSHIFT((w & 0xFFU), 8) | RSHIFT((w & 0xFF00U), 8)); }
 
-/*!\brief Swap endians of the contents of a 32b value
+/*!\brief Swap endians of the contents of a 32b value (little -> big, big -> little)
 ** \param[in] d - 32b value
 ** \return Swapped value
 **/
 __INLINE DWORD SWAP_END32B(const DWORD d) {
 	return (DWORD) (LSHIFT(SWAP_END16B(d & 0xFFFFUL), 16) | SWAP_END16B(RSHIFT((d & 0xFFFF0000UL), 16))); }
 
-/*!\brief Swap endians of the contents of a 64b value
+/*!\brief Swap endians of the contents of a 64b value (little -> big, big -> little)
 ** \param[in] l - 64b value
 ** \return Swapped value
 **/
@@ -190,21 +226,21 @@ __INLINE LWORD SWAP_END64B(const LWORD l) {
 	return (LWORD) (LSHIFT64(SWAP_END32B(l & 0xFFFFFFFFULL), 32) | SWAP_END32B(RSHIFT64((l & 0xFFFFFFFF00000000ULL), 32))); }
 
 
-/*!\brief Swap endians of a 16b tab
+/*!\brief Swap endians of a 16b tab (little -> big, big -> little)
 ** \param[in] tab - tab of 16b values
 ** \param[in] nb - nb of values in tab
 **/
 __INLINE void INLINE__ SWAP_END16B_TAB(WORD tab[], const WORD nb) {
 	for (unsigned int i = 0 ; i < nb ; i++)	tab[i] = SWAP_END16B(tab[i]); }
 
-/*!\brief Swap endians of a 32b tab
+/*!\brief Swap endians of a 32b tab (little -> big, big -> little)
 ** \param[in] tab - tab of 32b values
 ** \param[in] nb - nb of values in tab
 **/
 __INLINE void INLINE__ SWAP_END32B_TAB(DWORD tab[], const WORD nb) {
 	for (unsigned int i = 0 ; i < nb ; i++)	tab[i] = SWAP_END32B(tab[i]); }
 
-/*!\brief Swap endians of a 64b tab
+/*!\brief Swap endians of a 64b tab (little -> big, big -> little)
 ** \param[in] tab - tab of 64b values
 ** \param[in] nb - nb of values in tab
 **/
@@ -242,8 +278,9 @@ __INLINE bool INLINE__ inRange(const SDWORD val, const SDWORD low, const SDWORD 
 /************************/
 
 /*!\brief Get floating point number decimal part
-** \note in need to print floats, add '-u _printf_float' in Linker options
-** \warning enabling floating point support from linker seems to fubar printing long variables
+** \note in need to print floats, add '-u _printf_float' in Linker options.
+** \warning enabling floating point support from linker seems to fubar printing long variables,
+** 			so welcome get_fp_dec for the purpose.
 ** \param[in] f - floating point value
 ** \param[in] nb - Number of decimal to get after floating point
 ** \return nb decimal part as integer