+++ arm_typedefs.h
@@ -75,6 +75,17 @@ typedef enum eEdge {
 	Falling		//!< Falling edge
 } eEdge;
 
+/*!\enum eEndian
+** \brief Core endian
+**/
+typedef enum eEndian {
+	Endian_little = 0,		//!< Little endian configured MCU
+	Endian_big,				//!< Big endian configured MCU
+	Endian_mid_little,		//!< Middle little endian configured MCU (PDP-11)
+	Endian_mid_big,			//!< Middle big endian configured MCU (Honeywell 316)
+	Endian_unknown			//!< Unknown endian MCU
+} eEndian;
+
 
 /*!\struct StructBitfield8
 ** \brief Bitfield 8b